Minimize mobile WebView settings support
Inline the MobileAppWebView concern into Settings::BaseController, drop
the unused mobile_app_web_view? helper, and scope is_mobile_app_web_view
to the settings inertia_share instead of broadening global InertiaRendering.
